Refactored VHR table
